backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9

# Optional dependencies (install only if needed)
# openai==1.3.0
# azure-ai-formrecognizer==3.3.0
# google-generativeai==0.3.0